[Feature] Sorting blocks with drag and drop (#14)
This PR adds the functionality of using the grip button of a block to move that block around in order to reorder the blocks in a note. This functionality was implemented using a package based on [SortableJS](https://github.com/SortableJS/Sortable#options), called [sortablejs-vue3](https://github.com/MaxLeiter/sortablejs-vue3)

It also adds a small bug fix when converting a block into a checkbox where the 'selected' property would not be propertly created, thus failing the conversion.