[Feature] Move focus between blocks with arrow keys (#16)
# Main feature

Allowing the user to move the caret (cursor) between blocks with the arrow keys, which was not posible before. 

While this behaviour itself doesn't seem to be buggy, I've had to change the default behaviour of how the HTML Text nodes are parsed and render under different situations (on "regular" render, when pressing Shift + Enter, when pressing backspace, etc.), so while the situations that I have covered should function correctly, due to it being a big workload (I think that doing an editor with all of these functions nailed down could very well be a Final Degree Project in itself) I've left out other types of interactions like pressing the "Delete" button, or when selecting the whole content with Ctrl/Command + A and then deleting it, etc. 

This means that in those scenarios they won't behave like they should. For example, deleting the whole content of a block with Delete will cause the caret to appear next to the placeholder instead of it. In Chromium browsers and Safari this rarely happens, but in Firefox it is pretty much guaranteed.

# Other things

* I have also refactored all the blocks to instead have an inner dynamic component called `ContentEditableComponent` which implements all the functions related to the contentEditabe, since they were repeated over all the blocks.
* With this, I have also separated the now huge `useTextBasedBlock` composable into several smaller and more concise composables: `useHandleArrowKeys`, `useHandleBackspace`, `useRenderBlockContent`, `useHandleShiftEnter`, `useIsContentEmpty`.
* Bug fixes for the note's title.
* Minor bugfixes all over the place.