[Feature] Implement auth frontend side (#18)
By using [auth0](https://auth0.com/) I've implemented authorisation with a more frontend(ish) approach, since this project is mainly focused on the frontend. 
This is done by using their Vue 3 SDK, which allows the creations of navigation guards and functions that redirect the user an auth0 hosted page that, on login, sends the frontend the information about the user in token form, which is automatically locally stored (and thus the session is also persisted) by the SDK. 

Then to access the resources stored in my own MongoDB the requests that are sent to the API are accompanied by an Access Token provided by auth0, which is then validated in the API with the package [`express-oauth2-jwt-bearer` ](https://www.npmjs.com/package/express-oauth2-jwt-bearer). 

While this is not as secure as doing everything from the backend and just using a cookie to keep the session alive, I believe it's more than enough for this project (specially better than doing my own email-password login system), much more simple than the former alternative and still quite secure. The main risk would be XSS attacks if there were any vulnerabilities in the browser or packages that I use (which I've tried to keep to a minimum). If deployed the only consideration would be to make sure to use HTTPS to encrypt the token when sent.