Fix: fixed click issue on `MoveItemDialog`
Fixed bug where folder would not be selected as a new parent if the user clicked in the empty space allocated for the deployable button for folders that have no children.

Added `tabindex` to folders to make them navigable via `Tab` key.

Added key listener for the `space` key to `MoveItemDialogTargetFolder` to improve keyboard accesibility.