Changed navigation rendering condition
Changed rendering of sidebar items from v-if to v-show. This way they are rendered from the start and won't be unmounted when hidden. This allows to keep the status of whether a folder is open or not in deeply nested hierarchies. It will also be faster when opening/closing, though it will require more time for the first load if there are lots of sidebar items