[Feature] Add breadcrumbs to header (#24)
* Added a Breadcrumbs set of components to the header to quickly show the user the file hierarchy whenever they are editing a file.
* Refactored the Save button to its own component to cleanup the header now that it contains more elements.
* Changed the `v-if` rendering condition of a `NavigationFolder`'s contents to a `v-show`. 
The `v-if` caused its sub-elements to be unmounted from the DOM when hidden and only mounted again when shown. This would be good in a situation where there are lots of folders, subfolders and notes, since this would avoid loading them all from the start and then keeping them in the DOM memory, unlike with `v-show`, making it a bit more lightweight. However, without further logic this would mean that if you had a deeply nested set of folders, whenever you closed a parent folder, its child folders would get their open/closed state set back to closed. 
Since this is a POC/MVP, I considered that the convenience of closing-opening a hierarchy of folders and keeping the open/closed state of the subfolders was more relevant, since in this project it'll be very rare to reach that many folders as to cause a performance problem.
* Fixes for all the Dialogs, which would wrongly show the message variable when disappearing instead of keeping the text.